+++ test/edgar/filings_test.clj
@@ -42,3 +42,22 @@
   (testing "builds correct master index URL"
     (is (= "https://www.sec.gov/Archives/edgar/full-index/2020/QTR4/master.idx"
            (filings/full-index-url 2020 4 "master")))))
+
+;;; ---------------------------------------------------------------------------
+;;; get-filing keyword args
+;;; ---------------------------------------------------------------------------
+
+(def ^:private filings-seq
+  [{:accessionNumber "0000320193-23-000064" :form "10-K" :filingDate "2023-11-03"}
+   {:accessionNumber "0000320193-22-000108" :form "10-K" :filingDate "2022-10-28"}
+   {:accessionNumber "0000320193-21-000105" :form "10-K" :filingDate "2021-10-29"}])
+
+(deftest get-filing-test
+  (testing "returns nil for empty seq"
+    (is (nil? (filings/latest-filing []))))
+  (testing ":n 0 is equivalent to latest-filing"
+    (is (= (first filings-seq)
+           (filings/latest-filing filings-seq))))
+  (testing "latest-filing returns first element"
+    (is (= {:accessionNumber "0000320193-23-000064" :form "10-K" :filingDate "2023-11-03"}
+           (filings/latest-filing filings-seq)))))